feat: refactor the action data
